Tests: When checking compilation of extension modules, check if CPython can import them.
* For Python3.6 with debug and non-debug installed, there are modules
  that fail to import. Don't try those with Nuitka as they also are
  going to fail, which is not an issue.

* Also when checking for syntax errors, use the intended Python, not the
  one the runner is using itself.

+++ tests/library/compile_extension_modules.py
@@ -38,7 +38,8 @@ from nuitka.tools.testing.Common import (
     my_print,
     createSearchMode,
     compileLibraryTest,
-    check_output
+    check_output,
+    checkSucceedsWithCPython
 )
 
 setup(needs_io_encoding = True)
@@ -103,16 +104,21 @@ def action(stage_dir, root, path):
 
     command.append(filename)
 
-    try:
-        output = check_output(command).splitlines()
-    except Exception:
-        raise
-    else:
-        if output[-1] != b"OK":
-            sys.exit("FAIL")
-        my_print("OK")
+    if checkSucceedsWithCPython(filename):
+        try:
+            output = check_output(command).splitlines()
+        except Exception:
+            raise
+        else:
+            if output[-1] != b"OK":
 
-        shutil.rmtree(filename[:-3] + ".dist")
+
+                sys.exit("FAIL")
+            my_print("OK")
+
+            shutil.rmtree(filename[:-3] + ".dist")
+    else:
+        my_print("SKIP (does not work with CPython)")
 
 
 compileLibraryTest(
